Document-processor-AI/requirements.txt
langchain
langchain-openai
langchain-community
openai>=1.0.0
python-dotenv
redis==5.0.4
redisvl
langchain_redis
prometheus_client
boto3
fastapi
uvicorn[standard]
pypdf
numpy